fix(book): 修复图书管理中的ISBN验证和数据类型问题
- 在BookController中添加ISBN非空验证并抛出BaseException异常
- 修正图书更新逻辑，确保同一本图书可以更新而不报重复错误
- 将BookDTO中id字段类型从Integer改为Long以匹配实际数据库主键
- 为BookPageQueryDTO中的page和pageSize字段设置默认值避免空值问题
- 在BookServiceImpl中添加新增图书时的ISBN非空验证
- 添加BOOK_ISBN_NULL常量到MessageConstant中用于错误提示
- 配置WebMvcConfiguration允许管理员登出接口不经过JWT拦截器验证

+++ store-pojo/src/main/java/com/PTU/dto/BookDTO.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Data
 public class BookDTO {
     @TableId(type = IdType.AUTO)
-    private Integer id; // 主键ID
+    private Long id; // 主键ID
     private String title; // 书名
     private String author; // 作者
     private String publisher; // 出版社
